requirements.txt
# --- CORE API & LLM ---
google-genai>=1.65.0          
langchain-google-genai>=4.2.2 

# --- PUBMED DATA INGESTION ---
biopython>=1.87              
requests>=2.33.1             

# --- VECTOR STORAGE & PROCESSING ---
chromadb>=1.5.8              
numpy>=1.26.0                

# --- UTILS ---
python-dotenv>=1.2.2         
pydantic>=2.7.0
langchain>=0.3.0
langchain-community>=0.3.0